+++ pylinkage/interface/__init__.py
@@ -1,0 +1,11 @@
+"""
+Classes that represent the main interface with the library.
+"""
+
+from .linkage import (
+    Static,
+    Pivot,
+    Crank,
+    Fixed,
+    Linkage,
+)